feat: refactor network and cluster hooks for improved modularity and clarity
