+++ app/(main)/_lib/clusters.d.ts
@@ -1,0 +1,5 @@
+export interface ClusterGroup {
+  name: string;
+  description?: string;
+  members?: string[];
+}
