+++ app/(main)/_lib/networkIntegrations.d.ts
@@ -1,0 +1,5 @@
+export interface NetworkIntegration {
+  name: string;
+  description?: string;
+  type?: string;
+}
