+++ app/(main)/_lib/networkZones.d.ts
@@ -1,0 +1,4 @@
+export interface NetworkZone {
+  name: string;
+  description?: string;
+}
